[REFACTOR] #12 : BaseEntity 위치 수정
